Make it possible to use the detect without an external API
For local setup, it's easier to have all in the same API.

Added a configuration `DETECT_URL` to configure the detection api url.
If set to `""`, the detection will be done locally.
The default is still `http://localhost:8001` as before.

I'm not a fan of the default value, but I think it can be enough for the
moment.